Invoke the callbacks of localStorage async
This avoids issues with other libraries like Thorax that do not expect that the success or error callback are invoked synchronously.
It also ensures that the behaviour between onlineSync and offlineSync are equal. This avoids timing issues when migrating from offline only applications (prototyping) to applications with an online backend. Additionally it avoids issues with callbacks that perform time consuming operations like updating the dom. This would unnecessarily suspend the processing of the sync call.